refactor(ui): extract AssistantMessageDisplay + AssistantParts + ContextToolGroup
Slice 11b.1 step 6/14 of the message-part.tsx split. Moves the entire
legacy assistant rendering surface into a dedicated module:

  `assistant-message-display.tsx` (470 lines)
    - `AssistantMessageDisplay`  single-message dispatcher
    - `AssistantParts`           multi-message round renderer (used by
                                 the legacy SessionTurn shell)
    - `ContextToolGroup`         collapsible context-tool row
    - `ExaOutput`                links-only fallback for webfetch/search
    - private helpers: `urls`, `contextToolDetail`, `contextToolTrigger`,
      `contextToolSummary`

Back-compat preserved: `message-part.tsx` re-exports
`AssistantMessageDisplay` and `AssistantParts` so existing consumers
(session-turn.tsx imports `AssistantParts` directly) keep resolving;
`ExaOutput` is imported into message-part.tsx and used by the websearch
tool renderer that still lives here.

This is the larger of the two named extractions in design doc §1
("extracted from message-part.tsx L901"). The User display extraction
is the next step.

message-part.tsx 1986 → 1607 lines (-379). Typecheck GREEN, 478 pass /
3 baseline fail unchanged.
